:recycle: Fix lint
