+++ tagupy/__init__.py
@@ -1,6 +1,12 @@
 """Top-level package for TaguPy."""
+
 from .analysis import StatsAnalysis
 from .experiment import MatrixGenerator
+
+__all__ = [
+    StatsAnalysis,
+    MatrixGenerator,
+]
 
 __author__ = """Sei Takeda"""
 __email__ = 'sei06k14@gmail.com'
