+++ tagupy/analysis/__init__.py
@@ -1,1 +1,5 @@
 from .stats_analysis import StatsAnalysis
+
+__all__ = [
+    StatsAnalysis
+]
