+++ tagupy/experiment/__init__.py
@@ -1,1 +1,5 @@
 from .matrix_generator import MatrixGenerator
+
+__all__ = [
+    MatrixGenerator,
+]
