--- tagupy/tagupy.py
@@ -1,6 +1,0 @@
-"""
-Main module.
-
-TODO:
-    Write Experiment management conductor program.
-"""
