+++ tests/conftest.py
@@ -1,0 +1,1 @@
+collect_ignore = ['setup.py']
