+++ tests/test_cli.py
@@ -1,0 +1,18 @@
+#!/usr/bin/env python
+
+"""Tests for `tagupy` package."""
+
+from click.testing import CliRunner
+
+from tagupy import cli
+
+
+def test_command_line_interface():
+    """Test the CLI."""
+    runner = CliRunner()
+    result = runner.invoke(cli.main)
+    assert result.exit_code == 0
+    assert 'tagupy.cli.main' in result.output
+    help_result = runner.invoke(cli.main, ['--help'])
+    assert help_result.exit_code == 0
+    assert '--help  Show this message and exit.' in help_result.output
